Moved JsonController from version 1 to V1 directory
